websocket: move test hooks into helper file
