+++ exchange/websocket/manager.go
@@ -772,74 +772,6 @@ func (m *Manager) IsEnabled() bool {
 	return m.enabled.Load()
 }
 
-// SetSubscriptionsNotRequired configures the manager to connect without
-// generating default subscriptions. This is primarily used by test helpers
-// that need request-response websocket connectivity without stream bootstraps.
-// It is intended as a pre-connect test-mode mutation; tests that later need
-// the manager's normal default-subscription bootstrap should use a fresh
-// exchange/manager instance.
-func (m *Manager) SetSubscriptionsNotRequired() {
-	m.m.Lock()
-	defer m.m.Unlock()
-
-	if !m.useMultiConnectionManagement {
-		m.GenerateSubs = func() (subscription.List, error) { return subscription.List{}, nil }
-		return
-	}
-
-	for _, ws := range m.connectionManager {
-		if ws.setup == nil {
-			log.Warnf(log.WebsocketMgr, "%s websocket: missing connection setup while disabling required subscriptions; creating empty setup", m.exchangeName)
-			ws.setup = &ConnectionSetup{}
-		}
-		ws.setup.SubscriptionsNotRequired = true
-	}
-}
-
-// SetAllConnectionURLs configures every managed websocket connection to use the
-// same URL. This is primarily used by test helpers that need to redirect all
-// websocket traffic through a single mock server before connecting. Calling
-// this after Connect has started returns an error.
-func (m *Manager) SetAllConnectionURLs(u string) error {
-	if err := common.NilGuard(m); err != nil {
-		return err
-	}
-	if err := checkWebsocketURL(u); err != nil {
-		return err
-	}
-
-	m.m.Lock()
-	defer m.m.Unlock()
-
-	if m.IsConnecting() {
-		return fmt.Errorf("%v %w: SetAllConnectionURLs must be called before Connect", m.exchangeName, errAlreadyReconnecting)
-	}
-	if m.IsConnected() {
-		return fmt.Errorf("%v %w: SetAllConnectionURLs must be called before Connect", m.exchangeName, errAlreadyConnected)
-	}
-
-	if !m.useMultiConnectionManagement {
-		m.runningURL = u
-		m.runningURLAuth = u
-		if m.Conn != nil {
-			m.Conn.SetURL(u)
-		}
-		if m.AuthConn != nil {
-			m.AuthConn.SetURL(u)
-		}
-		return nil
-	}
-
-	for _, ws := range m.connectionManager {
-		if ws.setup == nil {
-			log.Warnf(log.WebsocketMgr, "%s websocket: missing connection setup while updating connection URLs; creating empty setup", m.exchangeName)
-			ws.setup = &ConnectionSetup{}
-		}
-		ws.setup.URL = u
-	}
-	return nil
-}
-
 // CanUseAuthenticatedWebsocketForWrapper Handles a common check to
 // verify whether a wrapper can use an authenticated websocket endpoint
 func (m *Manager) CanUseAuthenticatedWebsocketForWrapper() bool {
